+++ app/models/carto/user.rb
@@ -220,12 +220,12 @@ class Carto::User < ActiveRecord::Base
   end
 
   def twitter_datasource_enabled
-    read_attribute(:twitter_datasource_enabled) && twitter_configured?
+    (read_attribute(:twitter_datasource_enabled) || organization.try(&:twitter_datasource_enabled)) && twitter_configured?
   end
 
   def twitter_configured?
-    config = CartoDB::Datasources::DatasourcesFactory.config_for(Search::Twitter::DATASOURCE_NAME, self)
-    !config['username'].empty? && !config['password'].empty?
+    # DatasourcesFactory.config_for takes configuration from organization if user is an organization user
+    CartoDB::Datasources::DatasourcesFactory.customized_config?(Search::Twitter::DATASOURCE_NAME, self)
   end
 
   # TODO: this is the correct name for what's stored in the model, refactor changing that name